renamed the ml_model directory to ml
