+++ main.py
@@ -2,8 +2,8 @@ import sys
 import torch
 from torch.utils.data import DataLoader
 from torch import nn
-from ml_model.dataset import CustomDataset
-from ml_model.model import NeuralNetwork
+from ml.dataset import CustomDataset
+from ml.model import NeuralNetwork
 
 BATCH_SIZE = 64
